simplify the code on cartReducer
